% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{boot_auc_folds}
\alias{boot_auc_folds}
\title{Compute bootstraps of the mean ROC}
\usage{
boot_auc_folds(predictions, folds, targets, num_samples = 1000,
  alt_predictions = NULL, fixed_axis = c("sensitivity", "specificity"),
  points = seq(0, 1, by = 0.01), stratified = T)
}
\arguments{
\item{predictions}{sample predictions}

\item{folds}{vector of fold classification for each sample}

\item{targets}{target values}

\item{num_samples}{number of bootstrap samples to take}

\item{alt_predictions}{alternative predictions to compare}

\item{fixed_axis}{whether to fix sensitivity or specificity in the ROC curves}

\item{points}{number of points to calculate for ROC curves}
}
\value{
bootstrapped ROC curves
}
\description{
Compute the ROC curve separately for each fold, and compute the mean
ROC curve, instead of pooling predictions.
}
