% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{calc_mean_roc}
\alias{calc_mean_roc}
\title{Calculate the mean ROC curve from a list of predictors and targets}
\usage{
calc_mean_roc(predictions_list, targets_list, points,
  fixed_axis = c("specificity", "sensitivity"), boot_index = NULL,
  return_folds = FALSE)
}
\arguments{
\item{predictions_list}{list of predictors}

\item{targets_list}{list of targets}

\item{points}{points to calculate the ROC curves at}

\item{fixed_axis}{whether \code{points} fixes sensitivity or specificity}

\item{boot_index}{list of bootstrap indices for each fold. Generated automatically if NULL}

\item{return_folds}{if TRUE, return the ROC curve and AUC for each fold as well as the mean}
}
\value{
list containing mean ROC curve and AUC
}
\description{
Calculate the mean ROC curve from a list of predictors and targets
}
