% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkFlowRate}
\alias{checkFlowRate}
\title{Check the minimum flow rates for all the runs in a dataset}
\usage{
checkFlowRate(dir, threshold = 1.9, filePattern = ".*[.](txt|asc)",
  moveTo = NULL)
}
\arguments{
\item{dir}{Directory pat to check}

\item{filePattern}{Regex pattern of exported FAIMS files}
}
\value{
data frame listing minimum flow rate per folder
}
\description{
Check the minimum flow rates for all the runs in a dataset
}

