% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{ClassifierModels}
\alias{ClassifierModels}
\title{Run a set of classification models on input training, test data sets}
\usage{
ClassifierModels(data.train, targetValues, data.test, models, kFolds = 2,
  repeats = 5, tuneLength = 5, verbose = F)
}
\arguments{
\item{data.train}{a (nItems * nFeatures) data frame}

\item{targetValues}{a logical vector}

\item{data.test}{a (nItems * nFeatures) data frame}
}
\value{
a data frame containing prediction probabilities for each classification algorithm.
These are the predicted probabililty of outcome==TRUE
}
\description{
Run a set of classification models on input training, test data sets
}

