% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{compare_auc_bootstrap}
\alias{compare_auc_bootstrap}
\title{Compute a bootstrap of the difference between two ROC curve AUCs}
\usage{
compare_auc_bootstrap(roc_1, roc_2, rep = 10000)
}
\arguments{
\item{roc_1}{first ROC curve (from pROC)}

\item{roc_2}{second ROC curve (from pROC)}

\item{rep}{number of bootstrap samples}
}
\description{
Compute a bootstrap of the difference between two ROC curve AUCs
}

