% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{Cross-validation for classification models}
\usage{
CrossValidation(data.train, targetValues, models = c("rf", "glmnet",
  "svmRadial", "svmLinear", "gbm", "nnet"), nFolds = 10, stratified = TRUE,
  SGoF = NULL, threshold = NULL, nKeep = NULL, verbose = FALSE,
  heatmap = FALSE, PCA = FALSE, extraData = NULL, tuneKFolds = 2,
  tuneRepeats = 5, tuneLength = 5, folds = NULL)
}
\arguments{
\item{extraData}{}
}
\description{
Cross-validation for classification models
}

