% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{CrossValRocCurves}
\alias{CrossValRocCurves}
\title{Compute ROC objects for the results of a cross validation}
\usage{
CrossValRocCurves(crossVal.obj)
}
\arguments{
\item{crossVal.obj}{the output from \link{CrossValidation}}
}
\value{
A list of \link{pROC::roc} objects and a summary of the results
}
\description{
Compute ROC objects for the results of a cross validation
}

