% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deleteFAIMSSample}
\alias{deleteFAIMSSample}
\title{Delete samples from a FAIMS object}
\usage{
deleteFAIMSSample(FAIMSObject, deleteIndices)
}
\arguments{
\item{FAIMSObject}{a FAIMS object}

\item{deleteIndices}{indices to delete}
}
\value{
a FAIMS object
}
\description{
Delete samples from a FAIMS object
}

