% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoiseFaimsData.localCorr}
\alias{denoiseFaimsData.localCorr}
\title{Remove background noise from FAIMS data using local correlations}
\usage{
denoiseFaimsData.localCorr(dataMatrix, neighbourhoodSizeCol = 1,
  neighbourhoodSizeRow = 1, alpha = 0.05, plot = TRUE)
}
\arguments{
\item{dataMatrix}{FAIMS data matrix}

\item{alpha}{p-value required for a pixel to be considered correlated to its neighbours}

\item{neighbourhoodSize}{the size of the neighbourhood in which to look for local correlations}
}
\value{
A FAIMS data matrix
}
\description{
The signal in FAIMS data exhibits a high degree of local correlation. This
function zeroes out a fraction of pixels with the lowest degree of local
correlation.
}

