% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoiseFaimsData}
\alias{denoiseFaimsData}
\title{Remove background noise from FAIMS data}
\usage{
denoiseFaimsData(faimsObject, fractionNoise = 0.05,
  fractionToRemove = 0.99)
}
\arguments{
\item{fractionNoise}{The percentage of pixels to identify as "noise"}

\item{fractionToRemove}{The quantile value of the noise pixels to subtract from
all pixels}

\item{FAIMSObject}{FAIMS object}
}
\value{
a denoised FAIMS data matrix
}
\description{
Denoises FAIMS data. First identifies pixels likely to be solely noise (mean
value across all samples close to zero). Then identifies a value to subtract
from all pixels by taking the fractionToRemove quantile of all the values in
the noise pixels, and subtracts this from the absolute value of each pixels,
zeroing any values which are negative after subtracting background noise.
}
