% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{perform a feature selection using a Wilcoxon rank-sum test}
\usage{
FeatureSelection(dataMatrix, targetValues)
}
\arguments{
\item{nKeep}{}
}
\description{
perform a feature selection using a Wilcoxon rank-sum test
}

