% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{findNoiseFaimsData.localCorr}
\alias{findNoiseFaimsData.localCorr}
\title{Remove background noise from FAIMS data using local correlations}
\usage{
findNoiseFaimsData.localCorr(FAIMSObject, neighbourhoodSizeCol = 1,
  neighbourhoodSizeRow = 1, fractionToRemove = 0.65, plot = TRUE)
}
\arguments{
\item{FAIMSObject}{a FAIMS object}

\item{neighbourhoodSize}{the size of the neighbourhood in which to look for local correlations}

\item{alpha}{p-value required for a pixel to be considered correlated to its neighbours}
}
\value{
Logical vector containing TRUE when columns are noise
}
\description{
The signal in FAIMS data exhibits a high degree of local correlation. This
function zeroes out a fraction of pixels with the lowest degree of local
correlation.
}
