% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{findNoiseFaimsData.sd}
\alias{findNoiseFaimsData.sd}
\title{Identify background noise from FAIMS data}
\usage{
findNoiseFaimsData.sd(faimsObject, fractionNoise = 0.65, plot = TRUE)
}
\arguments{
\item{fractionNoise}{The percentage of pixels to identify as "noise"}

\item{FAIMSObject}{FAIMS object}
}
\value{
Logical vector containing TRUE when columns are noise
}
\description{
Identifies pixels likely to be solely noise based on standard deviation.
}
