% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{getNeighbourIndices}
\alias{getNeighbourIndices}
\title{Get Neighbour Indices}
\usage{
getNeighbourIndices(index, width, height, faimsDim)
}
\arguments{
\item{index}{index of the cell in the vector}

\item{width}{width of neighbourhood}

\item{height}{height of neighbourhood}

\item{faimsDim}{FAIMS matrix dimensions}
}
\value{
a vector of neighbour indices
}
\description{
Return the vector indices of a matrix cell's neighbours
}
