% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{getRocCurve}
\alias{getRocCurve}
\title{Return a ROC curve object, given input class probabilities}
\usage{
getRocCurve(predictions, targetValues, titleString = "", ci = FALSE,
  lr = FALSE, plotCurve = FALSE)
}
\arguments{
\item{ci}{Compute confidence interval?}

\item{lr}{}
}
\description{
Return a ROC curve object, given input class probabilities
}

