% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{GPLVMCrossValidation}
\alias{GPLVMCrossValidation}
\title{Title}
\usage{
GPLVMCrossValidation(data.train, targetValues, models = c("rf", "glmnet",
  "svmRadial", "svmLinear", "gbm", "nnet"), nFolds = 10, stratified = TRUE,
  sigma_d = 10^(c(-5, -3, -1, 0)), q = 3, verbose = FALSE,
  heatmap = FALSE, PCA = TRUE, extraData = NULL, tuneKFolds = 2,
  tuneRepeats = 5, tuneLength = 5, ...)
}
\arguments{
\item{tuneLength}{}
}
\description{
Title
}

