% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{knnMeanDistance}
\alias{knnMeanDistance}
\title{Find mean k-nearest neighbour distance for each sample}
\usage{
knnMeanDistance(dataMatrix, max.k = 10)
}
\arguments{
\item{max.k}{}
}
\description{
Find mean k-nearest neighbour distance for each sample
}

