% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{nKeep}
\alias{nKeep}
\title{Select the top n features}
\usage{
nKeep(scores, nKeep)
}
\arguments{
\item{scores}{a list of feature p-values}

\item{nKeep}{the number to keep}
}
\value{
A list if feature indices
}
\description{
Select the top n features
}
