% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{plotFAIMSdata}
\alias{plotFAIMSdata}
\title{Plot FAIMS data matrices}
\usage{
plotFAIMSdata(FAIMSObject, rowsToPlot, plotLayout = 1, absolute = T, ...)
}
\arguments{
\item{rowsToPlot}{the rows to plot}

\item{plotLayout}{a matrix to specify the layout for multiple plots per page (default is one plot per page)}

\item{absolute}{whether to plot the absolute value of the FAIMS data.}

\item{...}{additional parameters to pass to \code{image}}

\item{dataMatrix}{a matrix of FAIMS data}
}
\description{
Plot FAIMS data matrices
}
