% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{plotRocCurve}
\alias{plotRocCurve}
\title{Plot a ROC curve with CI}
\usage{
plotRocCurve(rocCurve, titleString = "", ci = TRUE)
}
\arguments{
\item{ciobj}{}
}
\value{

}
\description{
Plot a ROC curve with CI
}

