% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{prettyFAIMSPlot}
\alias{prettyFAIMSPlot}
\title{Plot FAIMS data matrices with axes and ion current scale}
\usage{
prettyFAIMSPlot(FAIMSObject, rowToPlot, runToPlot = 1, title = "")
}
\arguments{
\item{rowToPlot}{the row to plot}

\item{runToPlot}{the run to plot}

\item{dataMatrix}{a matrix of FAIMS data}
}
\description{
Plot FAIMS data matrices with axes and ion current scale
}
