% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progBarInit}
\alias{progBarInit}
\title{Initialise a progress bar object}
\usage{
progBarInit(range)
}
\arguments{
\item{range}{the range of values to cover}
}
\value{
a progress bar object
}
\description{
Initialise a progress bar object
}
