% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progBarUpdate}
\alias{progBarUpdate}
\title{Update a progress bar object}
\usage{
progBarUpdate(bar.obj, new.value)
}
\arguments{
\item{a}{progress bar object}

\item{the}{value to update the bar to}
}
\value{
an update progress bar object
}
\description{
Update a progress bar object
}
