% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadInFaimsData}
\alias{ReadInFaimsData}
\title{Read in FAIMS data from a list of files}
\usage{
ReadInFaimsData(fileList, dec = ".", minFlowRate = 1.9)
}
\arguments{
\item{fileList}{A list of files containing FAIMS data (in ASCII format)}

\item{dec}{Decimal symbol}

\item{minFlowRate}{minimum acceptable flow rate. If min flow rate falls below
this threshold an error is thrown}
}
\value{
A FAIMS object
}
\description{
Read in FAIMS data from a list of files
}
