% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadInFaimsDirectoriesAutosampler}
\alias{ReadInFaimsDirectoriesAutosampler}
\title{Read FAIMS data generated using an autosampler}
\usage{
ReadInFaimsDirectoriesAutosampler(dataPath, numRuns,
  filePattern = ".*[.](txt|asc)", runNumPattern = "[0-9]+[.](txt|asc)",
  dec = ".", minFlowRate = 1.9, dirPattern = ".*")
}
\arguments{
\item{dataPath}{The directory to scan for data}

\item{numRuns}{The number of runs to take}

\item{filePattern}{a regex which should match the names of the files you want to ingest}

\item{runNumPattern}{a regex which should match the run number of your data.
Any characters found will be stripped and the result converted using as.numeric}

\item{dec}{Decimal symbol}

\item{minFlowRate}{minimum acceptable flow rate. If min flow rate falls below
this threshold an error is thrown}
}
\value{
A FAIMS object
}
\description{
Read FAIMS data generated using an autosampler
}
