% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAIMS.R
\name{runFAIMS}
\alias{runFAIMS}
\title{Run a standard analysis given a FAIMS object and class labels}
\usage{
runFAIMS(FAIMSObject, targetValues, models = c("rf", "glmnet", "svmRadial",
  "svmLinear", "gbm", "nnet", "glm"))
}
\arguments{
\item{FAIMSObject}{a FAIMS object}

\item{targetValues}{class labels}

\item{models}{a list of \link{caret::train} models}
}
\value{
A list of results (see out$bestModelSummary and
  out$modelSelectSummary for a summary of results)
}
\description{
Run a standard analysis given a FAIMS object and class labels
}

