% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{runFold}
\alias{runFold}
\title{Utility function to create predictions for a given fold and set of variables}
\usage{
runFold(trainingData, trainingTargets, testData, models, tuneKFolds,
  tuneRepeats, tuneLength, keep, PCA, verbose, heatmap, extraTrainingData,
  extraTestData)
}
\arguments{
\item{heatmap}{}
}
\value{
predictions for this fold
}
\description{
Utility function to create predictions for a given fold and set of variables
}

