% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{SGoF}
\alias{SGoF}
\title{Significant Goodness of Fit}
\usage{
SGoF(scores, alpha)
}
\arguments{
\item{scores}{a list of feature p-values}

\item{alpha}{the significance level for the test}
}
\value{
The selected features' indices
}
\description{
Select features using the Significant Goodness of Fit meta-test
}
