% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{threshold}
\alias{threshold}
\title{Select features by threshold}
\usage{
threshold(scores, threshold)
}
\arguments{
\item{scores}{a list of feature p-values}

\item{threshold}{p-value threshold}
}
\value{
A list of feature indices
}
\description{
Select features whose score exceeds a given threshold
}
