% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletTransform.R
\name{WaveletTransform}
\alias{WaveletTransform}
\title{Perform a 1D wavelet transform on a FAIMS data matrix}
\usage{
WaveletTransform(FAIMSObject, discardTopLevels = 2)
}
\arguments{
\item{FAIMSObject}{a FAIMS object}

\item{discardTopLevels}{number of levels to discard as noise (0 keeps all
data)}
}
\value{
a matrix of wavelet-transformed FAIMS data
}
\description{
Perform a 1D wavelet transform on a FAIMS data matrix
}
