% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{boot_auc_folds}
\alias{boot_auc_folds}
\title{Compute bootstraps of the mean ROC}
\usage{
boot_auc_folds(predictions, folds, targets, num_samples = 1000,
  alt_predictions = NULL, fixed_axis = c("sensitivity", "specificity"),
  points = seq(0, 1, by = 0.01), stratified = T)
}
\arguments{
\item{points}{}
}
\description{
Compute the ROC curve separately for each fold, and compute the mean
ROC curve, instead of pooling predictions.
}
