% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{calc_mean_roc}
\alias{calc_mean_roc}
\title{Title}
\usage{
calc_mean_roc(predictions_list, targets_list, points, fixed_axis,
  boot_index = NULL, return_folds = FALSE)
}
\arguments{
\item{boot_index}{}
}
\description{
Title
}

