% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read (matt-ThinkPad-T430's conflicted copy 2016-11-15).R, R/read.R
\name{convertToArray}
\alias{convertToArray}
\title{Convert FAIMS data to an array}
\usage{
convertToArray(data)

convertToArray(data)
}
\arguments{
\item{data}{Faims data matrix}

\item{data}{Faims data matrix}
}
\description{
Convert FAIMS data to an array

Convert FAIMS data to an array
}

