% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{perform a feature selection using a Wilcoxon rank-sum test}
\usage{
FeatureSelection(dataMatrix, targetValues)
}
\arguments{
\item{dataMatrix}{a numeric matrix of training data}

\item{targetValues}{class labels (must be a logical vector)}
}
\value{
a list of scores, one for each sample. Lower is better.
}
\description{
perform a feature selection using a Wilcoxon rank-sum test
}

