% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{getRocCurve}
\alias{getRocCurve}
\title{Return a ROC curve object, given input class probabilities}
\usage{
getRocCurve(predictions, targetValues, titleString = "", ci = FALSE,
  lr = FALSE, plotCurve = FALSE)
}
\arguments{
\item{predictions}{vector of predictions}

\item{targetValues}{vector of class labels}

\item{titleString}{title string if \code{plotCurve==TRUE}}

\item{ci}{logical - compute confidence interval?}

\item{lr}{logical - compute likelihood ratios?}

\item{plotCurve}{logical - plot the resulting ROC curve?}
}
\value{
a \link{pROC::roc} object
}
\description{
Return a ROC curve object, given input class probabilities
}
