% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{knnMeanDistance}
\alias{knnMeanDistance}
\title{Find mean k-nearest neighbour distance for each sample}
\usage{
knnMeanDistance(dataMatrix, max.k = 10)
}
\arguments{
\item{dataMatrix}{a data matrix}

\item{max.k}{max k for mean k-nn distance}
}
\value{
a \code{nrow(dataMatrix)} by \code{k} matrix with the mean k-nn
  distance for each sample, with k along the columns and samples along the
  rows
}
\description{
Compute mean k-nn distances for outlier detection
}
