% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{plotRocCurve}
\alias{plotRocCurve}
\title{Plot a ROC curve with CI}
\usage{
plotRocCurve(rocCurve, titleString = "", ci = TRUE)
}
\arguments{
\item{rocCurve}{A ROC object}

\item{titleString}{Title for plot}

\item{ci}{Plot confidence intervals if TRUE}
}
\description{
Plot a ROC curve with CI
}

