% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read (matt-ThinkPad-T430's conflicted copy 2016-11-15).R, R/read.R
\name{ReadInFaimsData}
\alias{ReadInFaimsData}
\title{Read in FAIMS data from a list of files}
\usage{
ReadInFaimsData(fileList, dec = ".", minFlowRate = 1.9)

ReadInFaimsData(fileList, dec = ".", minFlowRate = 1.9)
}
\arguments{
\item{fileList}{A list of files containing FAIMS data (in ASCII format)}

\item{fileList}{A list of files containing FAIMS data (in ASCII format)}
}
\value{
a data matrix containing the FAIMS data (with one row per file)

a data matrix containing the FAIMS data (with one row per file)
}
\description{
Read in FAIMS data from a list of files

Read in FAIMS data from a list of files
}

