% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAIMS.R
\name{runFAIMS}
\alias{runFAIMS}
\title{Run a standard analysis given a FAIMS object and class labels}
\usage{
runFAIMS(FAIMSObject, targetValues, models = c("rf", "glmnet", "svmRadial",
  "svmLinear", "gbm", "nnet", "glm"), modelSelectFolds = NULL,
  modelSelectScores = NULL, bestModelFolds = NULL, bestModelScores = NULL,
  waveletData = NULL, SGoF = TRUE, nKeep = TRUE, extraData = NULL)
}
\arguments{
\item{FAIMSObject}{a FAIMS object}

\item{targetValues}{class labels}

\item{models}{a list of \link{caret::train} models}

\item{modelSelectFolds}{pre-generated folds for model selection}

\item{modelSelectScores}{pre-generated scores for model selection}

\item{bestModelFolds}{pre-generated folds for best model assessment}

\item{bestModelScores}{pre-generated scores for best model assessment}

\item{waveletData}{pre-computed wavelet data}

\item{SGoF}{Select variables using sequential goodness of fit? (only for PCA
analysis)}

\item{nKeep}{Select variables using keep top N?}

\item{extraData}{Additional data to feed to the classifier}
}
\value{
A list of results (see out$bestModelSummary and
  out$modelSelectSummary for a summary of results)
}
\description{
Run a standard analysis given a FAIMS object and class labels
}
