% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{runFold}
\alias{runFold}
\title{Utility function to create predictions for a given fold and set of variables}
\usage{
runFold(trainingData, trainingTargets, testData, models, tuneKFolds,
  tuneRepeats, tuneLength, keep, PCA, verbose, heatmap, extraTrainingData,
  extraTestData)
}
\arguments{
\item{trainingData}{A data frame of training data}

\item{trainingTargets}{A logical vector of training class labels}

\item{testData}{A data frame of test data}

\item{models}{A vector of \link{caret::train} models}

\item{tuneKFolds}{number of folds for parameter tuning}

\item{tuneRepeats}{number of repeats for parameter tuning}

\item{tuneLength}{number of parameters to try when tuning}

\item{keep}{numeric vector of column indices to keep in training and test data}

\item{PCA}{apply PCA?}

\item{verbose}{verbose output?}

\item{heatmap}{plot a heatmap before training models?}

\item{extraTrainingData}{additional training data}

\item{extraTestData}{additional test data}
}
\value{
predictions for this fold
}
\description{
Utility function to create predictions for a given fold and set of variables
}

