% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletTransform.R
\name{WaveletTransform}
\alias{WaveletTransform}
\title{Perform a 1D wavelet transform on a FAIMS data matrix}
\usage{
WaveletTransform(dataMatrix, discardTopLevels = 1)
}
\arguments{
\item{dataMatrix}{FAIMS data}
}
\value{
a matrix of wavelet-transformed FAIMS data
}
\description{
Perform a 1D wavelet transform on a FAIMS data matrix
}

