% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletTransform.R
\name{WaveletTransform_2D}
\alias{WaveletTransform_2D}
\title{Perform a 2D wavelet transform on a FAIMS data matrix}
\usage{
WaveletTransform_2D(dataMatrix, cropSize = NULL, discardHighestNLevels = 0)
}
\arguments{
\item{cropSize}{}
}
\description{
This function expects an input matrix of dimension (nDataItems * nFeatures)
each row is therefore the 1D representation of the 2D data for a single item
This means that this function needs to know the underlying dimensionality of the 2D FAIMS run
}

